Move templates out of source tree
