+++ src/sciwyrm/templates.py
@@ -1,6 +1,6 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2024 SciCat Project (https://github.com/SciCatProject/sciwyrm)
-"""Asset loaders for SciWyrm."""
+"""Template loading."""
 
 from functools import lru_cache
 from pathlib import Path
@@ -10,7 +10,7 @@ from fastapi import Depends
 from fastapi.templating import Jinja2Templates
 from jinja2 import Environment, FileSystemLoader
 
-from ..config import AppConfig, app_config
+from .config import AppConfig, app_config
 
 
 def get_templates(config: Annotated[AppConfig, Depends(app_config)]) -> Jinja2Templates:
@@ -18,10 +18,28 @@ def get_templates(config: Annotated[AppConfig, Depends(app_config)]) -> Jinja2Te
     return _make_template_handler(config.template_dir)
 
 
+def notebook_template_path(name: str, version: str) -> str:
+    """Return the relative path to a given notebook template.
+
+    Parameters
+    ----------
+    name:
+        Name of the template.
+    version:
+        Version of the template.
+
+    Returns
+    -------
+    :
+        The path to the template relative to the base template directory.
+    """
+    return f"notebook/{name}_v{version}.ipynb"
+
+
 @lru_cache(maxsize=1)
 def _make_template_handler(template_dir: Path) -> Jinja2Templates:
-    from .. import filters
-    from ..logging import get_logger
+    from . import filters
+    from .logging import get_logger
 
     get_logger().info("Loading templates from %s", template_dir)
     templates = Jinja2Templates(
